chore: rename fal_enc adapter to fal (#596)
