+++ adapter/src/dbt/adapters/fal/__init__.py
@@ -1,9 +1,8 @@
 from dbt.adapters.base import AdapterPlugin
 
-from dbt.adapters.fal.connections import FalCredentials
-from dbt.adapters.fal.impl import FalAdapter
+from dbt.adapters.fal.connections import FalEncCredentials
+from dbt.adapters.fal.impl import FalEncAdapter
 from dbt.include import fal
 
-Plugin = AdapterPlugin(
-    adapter=FalAdapter, credentials=FalCredentials, include_path=fal.PACKAGE_PATH
-)
+# NOTE: include_path has links to dbt-fal adapter's for Teleport
+Plugin = AdapterPlugin(adapter=FalEncAdapter, credentials=FalEncCredentials, include_path=fal.PACKAGE_PATH)
