+++ adapter/src/dbt/adapters/fal/connections.py
@@ -1,59 +1,18 @@
 from dataclasses import dataclass
-from typing import Optional
-import os
-
-from dbt.adapters.base import Credentials
-from dbt.dataclass_schema import StrEnum, ExtensibleDbtClassMixin
-
-from dbt.fal.adapters.python import PythonConnectionManager
-
-
-class TeleportTypeEnum(StrEnum):
-    LOCAL = "local"
-    REMOTE_S3 = "s3"
+from dbt.adapters.fal_experimental.connections import FalCredentials
 
 @dataclass
-class TeleportCredentials(ExtensibleDbtClassMixin):
-    type: TeleportTypeEnum
+class FalEncCredentials(FalCredentials):
 
-    # local
-    local_path: Optional[str] = os.getcwd()
+    db_profile: str = ''
 
-    # s3
-    s3_bucket: Optional[str] = None
-    s3_region: Optional[str] = None
-    s3_access_key_id: Optional[str] = None
-    s3_access_key: Optional[str] = None
-
-
-class FalConnectionManager(PythonConnectionManager):
-    TYPE = "fal"
-
-    @classmethod
-    def open(cls, connection):
-        raise NotImplementedError
-
-    def execute(self, compiled_code: str):
-        raise NotImplementedError
-
-    def cancel(self, connection):
-        raise NotImplementedError
-
-
-@dataclass
-class FalCredentials(Credentials):
-    default_environment: str = "local"
-
-    teleport: Optional[TeleportCredentials] = None
-
-    # NOTE: So we are allowed to not set them in profiles.yml
-    # they are ignored for now
-    database: str = ''
-    schema: str = ''
+    def _connection_keys(self):
+        return () + super()._connection_keys()
 
     @property
     def type(self):
         return "fal"
 
-    def _connection_keys(self):
-        return ()
+    @property
+    def unique_field(self):
+        return self.db_profile
