+++ adapter/src/dbt/adapters/fal_experimental/support/duckdb.py
@@ -1,6 +1,6 @@
 from dbt.adapters.base import BaseAdapter, BaseRelation
 from dbt.adapters.base.connections import AdapterResponse
-from dbt.adapters.fal.adapter_support import new_connection
+from dbt.adapters.fal_experimental.adapter_support import new_connection
 import pandas as pd
 from dbt.adapters.sql import SQLAdapter
 import duckdb
