+++ adapter/src/dbt/adapters/fal_experimental/teleport_adapter_support.py
@@ -1,5 +1,5 @@
 from dbt.adapters.base.impl import BaseAdapter
-from dbt.adapters.fal.connections import TeleportCredentials, TeleportTypeEnum
+from dbt.adapters.fal_experimental.connections import TeleportCredentials, TeleportTypeEnum
 
 from dbt.fal.adapters.teleport.impl import TeleportAdapter
 
@@ -10,11 +10,11 @@ def wrap_db_adapter(db_adapter: BaseAdapter, teleport_credentials: TeleportCrede
 
     # Wrap the adapter with a custom implementation
     if db_adapter.type() == 'duckdb':
-        import dbt.adapters.fal.teleport_support.duckdb as support_duckdb
+        import dbt.adapters.fal_experimental.teleport_support.duckdb as support_duckdb
         return support_duckdb.DuckDBAdapterTeleport(db_adapter, teleport_credentials)
 
     if db_adapter.type() == 'snowflake':
-        import dbt.adapters.fal.teleport_support.snowflake as support_snowflake
+        import dbt.adapters.fal_experimental.teleport_support.snowflake as support_snowflake
         if teleport_credentials.type == TeleportTypeEnum.REMOTE_S3:
             return support_snowflake.SnowflakeAdapterTeleport(db_adapter, teleport_credentials)
         else:
