+++ adapter/src/dbt/adapters/fal_experimental/utils.py
