+++ adapter/src/dbt/include/fal/macros/materializations/table.sql
@@ -1,6 +1,5 @@
-{% materialization table, adapter='fal', supported_languages=['python'] -%}
-
-  -- TODO: change to sync_teleport() when available in dbt-core
+{# HACK: materialization must be custom `fal_table` in order for `table` materializations to go into DB adapter #}
+{% materialization fal_table, adapter='fal', supported_languages=['python'] -%}
   {%- if adapter.is_teleport() -%}
     {%- for _ref in model.refs -%}
         {%- set resolved = ref(*_ref) -%}
@@ -17,7 +16,6 @@
   {%- endcall %}
 
   {{- return({'relations': [relation]}) }}
-
 {% endmaterialization %}
 
 {% macro py_write(code, relation) -%}
