--- adapter/src/dbt/include/fal_enc/macros/teleport_duckdb.sql
@@ -1,1 +1,0 @@
-../../fal/macros/teleport_duckdb.sql
