+++ adapter/src/dbt/include/fal_experimental/macros/teleport_duckdb.sql
@@ -1,0 +1,8 @@
+{% macro duckdb__copy_to(relation, url) -%}
+    COPY {{ relation }} TO '{{ url }}'
+{%- endmacro %}
+
+{% macro duckdb__copy_from_parquet(relation, url) -%}
+    CREATE OR REPLACE TABLE {{ relation }} AS
+        SELECT * FROM read_parquet('{{ url }}');
+{%- endmacro %}
