+++ adapter/src/dbt/include/fal_experimental/macros/teleport_snowflake.sql
